market_data/airflow/requirements.txt
psycopg2-binary
scrapy
apache-airflow-providers-google
pyarrow
apache-airflow-providers-postgres
requests
google-cloud-storage
google-cloud-bigquery